feat: rename steps folder (#23)
* feat: rename steps folder

* feat: apply same renaming to create-token folder
