+++ src/features/create-dao/CreateDAOModal.tsx
@@ -9,17 +9,20 @@ export interface CreateDAOModalProps extends BasicModalProps {
 	domainName: string;
 }
 
-export const CreateDAOModal: FC<CreateDAOModalProps> = ({ domainName, ...props }) => {
+export const CreateDAOModal: FC<CreateDAOModalProps> = ({
+	domainName,
+	...props
+}) => {
 	const [isOpen, setIsOpen] = useState(false);
 
 	const onClose = (): void => {
 		setIsOpen(false);
-	}
+	};
 
 	const onOpenChange = (open: boolean): void => {
 		if (!open) return;
 		setIsOpen(true);
-	}
+	};
 
 	return (
 		<Modal {...props} open={isOpen} onOpenChange={onOpenChange}>
