+++ src/features/create-dao/index.ts
@@ -4,4 +4,4 @@ export * from './CreateDAOForm';
 export * from './CreateDAOButton';
 export * from './CreateDAOModal';
 export * from './CreateDAO.types';
-export * from "./CreateDAOFormContext";
+export * from './CreateDAOFormContext';
