+++ src/features/create-token/Steps/DetailsForm.module.scss
