+++ src/features/create-token/Steps/DetailsForm.test.tsx
