+++ src/features/create-token/Steps/DetailsForm.tsx
